// src/generated-types/menu-item.ts
export interface MenuItem {
  id: string;
  slug: string;
  price: string;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  position: number;
  createdAt: Date | null;
  updatedAt: Date | null;
}

/** MenuItemWithTranslation represents a menu item with its translations. */
export interface MenuItemTranslation {
  id: string;
  title: string;
  description?: string | null | undefined;
  language: string;
}

/** MenuItemWithTranslation represents a menu item with its translations. */
export interface MenuItemWithTranslation {
  id: string;
  slug: string;
  price: string;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  position: number;
  translations: MenuItemTranslation[];
}

/** MenuItemListWithTranslation represents a list of menu items with their translations. */
export interface MenuItemListWithTranslation {
  menuItems: MenuItemWithTranslation[];
}

/** Request message for creating a menu item. */
export interface CreateMenuItemRequest {
  slug: string;
  price: string;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
  categoryId: string;
}

/** Request message for updating a menu item. */
export interface UpdateMenuItemRequest {
  id: string;
  slug?: string | null | undefined;
  price?: string | null | undefined;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
}

/** Request message for changing the position of a menu item. */
export interface ChangeMenuItemPositionRequest {
  id: string;
  position: number;
}

/** Request message for creating a menu item translation. */
export interface CreateMenuItemTranslationRequest {
  title: string;
  description?: string | null | undefined;
  language: string;
  itemId: string;
}

/** Request message for updating a menu item translation. */
export interface UpdateMenuItemTranslationRequest {
  id: string;
  title?: string | null | undefined;
  description?: string | null | undefined;
}

/** StatusResponse represents a standard response message indicating success or failure. */
export interface StatusResponse {
  success: boolean;
  message: string;
}